backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.26.0
